revert folder structure + use HeatingDevice as generic device
