+++ PyViCare/PyViCareHeatCurveCalculation.py
