+++ PyViCare/PyViCareHeatPump.py
@@ -1,6 +1,6 @@
 from typing import Any, List
 
-from PyViCare.heating.PyViCareHeatingDevice import (HeatingDevice, HeatingDeviceWithComponent)
+from PyViCare.PyViCareHeatingDevice import (HeatingDevice, HeatingDeviceWithComponent)
 from PyViCare.PyViCareUtils import handleNotSupported
